RF 00.00 | FolderTest | Moving 3 tests from ProjectFolder2Test to FolderTest, deleting ProjectFolder2Test
